==> 4 - casPratique.R <==
# Step 1 : Importation des données
# Creer une Dummy sur la variable de votre choix qui sera votre cible 
# La laguer de N periodes (pour faire une prévisions N-step ahead)
# Importer un set d'exogenes 
# Creer si besoin des indicateurs à partir de ces exogenes :
  # croissance, moyenne mobile, quantiles, ...
  # Indicateurs chartistes




# Step 2 : Creation de la base
# Fusionner les données
# S'adapter à la fréquence de la cible
# Gerer intelligement les NAs




# Step 3 : Modelisation
# Creer une base de train et une base de test 
# Modéliser la cible par le ou les modèles de votre choix
# Mesurer les performances (sensibilité, ppv, specificité, npv, prevalence,...)
# Optimiser votre modèle sur un ou plusieurs parametres
